Enhance get costcal for approval md

+++ Com.Danliris.Service.Sales.Lib/ViewModels/CostCalculationGarment/ProfitGarmentByComodityReportViewModel.cs
@@ -4,13 +4,13 @@ using System.Text;
 
 namespace Com.Danliris.Service.Sales.Lib.ViewModels.CostCalculationGarment
 {
-    public class ProfitGarmentByComodityReportViewModel 
+    public class ProfitGarmentByComodityReportViewModel
     {
         public int count { get; set; }
         public string ComodityCode { get; set; }
         public string ComodityName { get; set; }
-        public int Quantity { get; set; }
-        public string UOMUnit { get; set; }        
+        public double Quantity { get; set; }
+        public string UOMUnit { get; set; }
         public double Amount { get; set; }
         public double ProfitIDR { get; set; }
         public double ProfitUSD { get; set; }
